use size_t in EventPoll::get_num_events()

+++ src/EventPoll_epoll.cc
@@ -56,7 +56,7 @@ int EventPoll::poll(int timeout)
   return n;
 }
 
-int EventPoll::get_num_events()
+size_t EventPoll::get_num_events()
 {
   return num_events_;
 }
